Resolves #1022, reduces DB calls in writeCourseAssets()
Using a combination of a 'distinct' on the 'courseassets' collection
and an '$in' operator on the 'assets' collection we can determine the
assets for a given course using only two DB calls.

Previously the code was getting the 'courseassets' (including any
duplicated assets) and making a new request to the database in order
to get the details of the asset.